Add files via upload
When working with variables in factor models, sometimes the variability between scores can be quite disparate. To avoid warnings or to standardize the scale across all variables, we can use the POMP formula:

[ (actual score - minimum score) / (maximun score) - (minimum score) ] *100

This formula is easy to apply, but when dealing with a large number of variables, the process can slow down our analysis.

In this script, I provide a simple function designed to simplify the process. Both manual and automatic methods are included to demonstrate that both approaches yield similar results.