Add files via upload
Dealing with normal data is crucial for the success of any analysis. Sometimes, when working with a large number of variables, it can be challenging to identify deviations from normality. The no_normal() function can quickly detect variables that do not meet the expected skewness and/or kurtosis scores within the "normal" range.